test: move ketersediaan test to its component
